Suppress static leak lint warning for AsyncTask
See, this AsyncTask is *meant* to keep its parent instance
while it's running. There's nothing wrong with keeping
the instance for a short while.

If you think different, consider if it's really worth the extra
complexity of introducing a weak reference or a bus system or
something similar. Would that make things better or just fat?
Would that "better" even be noticeable to users?