Handle landscape to reverse landscape rotations
Restart the camera when the device is rotated directly from
landscape to reverse landscape orientation or vice versa.

The camera needs to be restarted because the camera orientation
was calculated for a certain orientation and rotating the
device by 180 degrees will result in an upside down camera
preview.

Note that onConfigurationChange() is not firing for those
orientation changes because the configuration did not change.