Add a rule to make an aar
While jitpack.io needs some time to sort things out - it
fails to build 1.7.0 because there are licenses that need
to be accepted - I'm considering going with aar's instead
of all that maven shit. Never quite liked it.

IMHO it's simply wrong to depend on something that's out
of your control.