Remove resetting touch listener in focusTo()
Because this would be an unexpected behaviour for callers of
focusTo().

The original idea was to stop listening as soon as there is a
RuntimeException to avoid future RuntimeExceptions for the
same cause.

This is now done in the touch listener in setTapToFocus()
and callers of focusTo() should do too.