Add setScaleType to control camera preview layout
On some devices, the camera frame has a different aspect ratio than
the screen.

Now, the camera preview can either be laid so that it completely
covers the available `View` area (which is `SCALE_TYPE_CENTER_CROP`,
the default), or so that it lies completely within it
(use `SCALE_TYPE_CENTER_INSIDE`).